+++ src/main/resources/intentionDescriptions/LuaCreateFunctionIntention/after.template.lua
@@ -1,0 +1,4 @@
+local function myFunc(arg1, arg2)
+end
+
+myFunc(1, 2)